Flyttar Mottaker-klassa over i modellen
